+++ Source/driver/Castro.cpp
@@ -47,6 +47,9 @@
 #endif
 
 #include <extern_parameters.H>
+#ifdef PROB_PARAMS
+#include <prob_parameters.H>
+#endif
 
 using namespace amrex;
 
@@ -131,6 +134,9 @@ Real         Castro::startCPUTime = 0.0;
 
 int          Castro::SDC_Source_Type = -1;
 int          Castro::num_state_type = 0;
+
+int          Castro::do_init_probparams = 0;
+
 
 namespace amrex {
     extern int compute_new_dt_on_regrid;
@@ -480,6 +486,14 @@ Castro::Castro (Amr&            papa,
     MultiFab::RegionTag amrlevel_tag("AmrLevel_Level_" + std::to_string(lev));
 
     buildMetrics();
+
+    // initialize the C++ values of the runtime parameters
+#ifdef PROB_PARAMS
+    if (do_init_probparams == 0) {
+      init_prob_parameters();
+      do_init_probparams = 1;
+    }
+#endif
 
     initMFs();
 
@@ -2013,13 +2027,11 @@ Castro::post_restart ()
             {
                 if (gravity->NoComposite() != 1)
                 {
-                   int use_previous_phi = 1;
-
                    // Update the maximum density, used in setting the solver tolerance.
 
                    gravity->update_max_rhs();
 
-                   gravity->multilevel_solve_for_new_phi(0,parent->finestLevel(),use_previous_phi);
+                   gravity->multilevel_solve_for_new_phi(0, parent->finestLevel());
                    if (gravity->test_results_of_solves() == 1)
                        gravity->test_composite_phi(level);
                 }
@@ -2218,15 +2230,13 @@ Castro::post_regrid (int lbase,
             if ( (level == lbase) && cur_time > 0.)
             {
                 if ( gravity->get_gravity_type() == "PoissonGrav" && (gravity->NoComposite() != 1) ) {
-                    int use_previous_phi = 1;
-
                     // Update the maximum density, used in setting the solver tolerance.
 
                     if (level == 0) {
                       gravity->update_max_rhs();
                     }
 
-                    gravity->multilevel_solve_for_new_phi(level,new_finest,use_previous_phi);
+                    gravity->multilevel_solve_for_new_phi(level, new_finest);
 
                 }
 
@@ -2658,7 +2668,15 @@ Castro::reflux(int crse_level, int fine_level)
             }
             for (int i = 0; i < BL_SPACEDIM; ++i) {
                 MultiFab::Add(*crse_lev.fluxes[i], *temp_fluxes[i], 0, 0, crse_lev.fluxes[i]->nComp(), 0);
-                MultiFab::Add(*crse_lev.mass_fluxes[i], *temp_fluxes[i], URHO, 0, 1, 0);
+
+                // The gravity and rotation source terms depend on the mass fluxes.
+                // These should be the same as the URHO component of the fluxes.
+                // This update must be a copy from the fluxes rather than an add
+                // from the flux register because the mass fluxes only represent
+                // the last subcycle of the previous timestep.
+
+                MultiFab::Copy(*crse_lev.mass_fluxes[i], *crse_lev.fluxes[i], URHO, 0, 1, 0);
+
                 temp_fluxes[i].reset();
             }
 
@@ -3057,6 +3075,52 @@ Castro::enforce_min_density (MultiFab& state_in, int ng)
 }
 
 void
+Castro::enforce_speed_limit (MultiFab& state_in, int ng)
+{
+    BL_PROFILE("Castro::enforce_speed_limit()");
+
+    // This routine sets the velocity in state_in to be no larger than the
+    // speed limit, if one has been applied.
+
+    if (castro::speed_limit <= 0.0_rt) return;
+
+#ifdef _OPENMP
+#pragma omp parallel
+#endif
+    for (MFIter mfi(state_in, TilingIfNotGPU()); mfi.isValid(); ++mfi) {
+
+        const Box& bx = mfi.growntilebox(ng);
+
+        auto u = state_in[mfi].array();
+
+        amrex::ParallelFor(bx,
+        [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k) noexcept
+        {
+            Real rho = u(i,j,k,URHO);
+            Real rhoInv = 1.0_rt / rho;
+
+            Real vx = u(i,j,k,UMX) * rhoInv;
+            Real vy = u(i,j,k,UMY) * rhoInv;
+            Real vz = u(i,j,k,UMZ) * rhoInv;
+
+            Real v = std::sqrt(vx * vx + vy * vy + vz * vz);
+
+            if (v > castro::speed_limit) {
+                Real reduce_factor = castro::speed_limit / v;
+
+                u(i,j,k,UMX) *= reduce_factor;
+                u(i,j,k,UMY) *= reduce_factor;
+                u(i,j,k,UMZ) *= reduce_factor;
+
+                u(i,j,k,UEDEN) -= 0.5_rt * rhoInv * (rho * vx * rho * vx - u(i,j,k,UMX) * u(i,j,k,UMX) +
+                                                     rho * vy * rho * vy - u(i,j,k,UMY) * u(i,j,k,UMY) +
+                                                     rho * vz * rho * vz - u(i,j,k,UMZ) * u(i,j,k,UMZ));
+            }
+        });
+    }
+}
+
+void
 Castro::avgDown (int state_indx)
 {
     BL_PROFILE("Castro::avgDown(state_indx)");
@@ -3200,6 +3264,8 @@ Castro::apply_tagging_func(TagBoxArray& tags, Real time, int j)
         const int8_t tagval   = (int8_t) TagBox::SET;
         const int8_t clearval = (int8_t) TagBox::CLEAR;
 
+        int lev = level;
+
         if (err_list_names[j] == "density") {
             amrex::ParallelFor(bx,
             [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
@@ -3208,7 +3274,7 @@ Castro::apply_tagging_func(TagBoxArray& tags, Real time, int j)
                             (int8_t*) AMREX_ARR4_TO_FORTRAN_ANYD(tagfab),
                             AMREX_ARR4_TO_FORTRAN_ANYD(datfab), ncomp,
                             AMREX_ZFILL(dx.data()), AMREX_ZFILL(problo.data()),
-                            tagval, clearval, time, level);
+                            tagval, clearval, time, lev);
             });
         }
         else if (err_list_names[j] == "Temp") {
@@ -3219,7 +3285,7 @@ Castro::apply_tagging_func(TagBoxArray& tags, Real time, int j)
                              (int8_t*) AMREX_ARR4_TO_FORTRAN_ANYD(tagfab),
                              AMREX_ARR4_TO_FORTRAN_ANYD(datfab), ncomp,
                              AMREX_ZFILL(dx.data()), AMREX_ZFILL(problo.data()),
-                             tagval, clearval, time, level);
+                             tagval, clearval, time, lev);
             });
         }
         else if (err_list_names[j] == "pressure") {
@@ -3230,7 +3296,7 @@ Castro::apply_tagging_func(TagBoxArray& tags, Real time, int j)
                               (int8_t*) AMREX_ARR4_TO_FORTRAN_ANYD(tagfab),
                               AMREX_ARR4_TO_FORTRAN_ANYD(datfab), ncomp,
                               AMREX_ZFILL(dx.data()), AMREX_ZFILL(problo.data()),
-                              tagval, clearval, time, level);
+                              tagval, clearval, time, lev);
             });
         }
         else if (err_list_names[j] == "x_velocity" || err_list_names[j] == "y_velocity" || err_list_names[j] == "z_velocity") {
@@ -3241,7 +3307,7 @@ Castro::apply_tagging_func(TagBoxArray& tags, Real time, int j)
                             (int8_t*) AMREX_ARR4_TO_FORTRAN_ANYD(tagfab),
                             AMREX_ARR4_TO_FORTRAN_ANYD(datfab), ncomp,
                             AMREX_ZFILL(dx.data()), AMREX_ZFILL(problo.data()),
-                            tagval, clearval, time, level);
+                            tagval, clearval, time, lev);
             });
         }
 #ifdef REACTIONS
@@ -3253,7 +3319,7 @@ Castro::apply_tagging_func(TagBoxArray& tags, Real time, int j)
                             (int8_t*) AMREX_ARR4_TO_FORTRAN_ANYD(tagfab),
                             AMREX_ARR4_TO_FORTRAN_ANYD(datfab), ncomp,
                             AMREX_ZFILL(dx.data()), AMREX_ZFILL(problo.data()),
-                            tagval, clearval, time, level);
+                            tagval, clearval, time, lev);
             });
         }
         else if (err_list_names[j] == "enuc") {
@@ -3264,7 +3330,7 @@ Castro::apply_tagging_func(TagBoxArray& tags, Real time, int j)
                              (int8_t*) AMREX_ARR4_TO_FORTRAN_ANYD(tagfab),
                              AMREX_ARR4_TO_FORTRAN_ANYD(datfab), ncomp,
                              AMREX_ZFILL(dx.data()), AMREX_ZFILL(problo.data()),
-                             tagval, clearval, time, level);
+                             tagval, clearval, time, lev);
             });
         }
 #endif
@@ -3277,7 +3343,7 @@ Castro::apply_tagging_func(TagBoxArray& tags, Real time, int j)
                             (int8_t*) AMREX_ARR4_TO_FORTRAN_ANYD(tagfab),
                             AMREX_ARR4_TO_FORTRAN_ANYD(datfab), ncomp,
                             AMREX_ZFILL(dx.data()), AMREX_ZFILL(problo.data()),
-                            tagval, clearval, time, level);
+                            tagval, clearval, time, lev);
             });
         }
 #endif
@@ -4191,6 +4257,10 @@ Castro::clean_state(
     // Enforce a minimum density.
 
     enforce_min_density(state_in, ng);
+
+    // Enforce the speed limit.
+
+    enforce_speed_limit(state_in, ng);
 
     // Ensure all species are normalized.